backend/requirements.txt
# DiffCOT Code Review Backend Dependencies

# Web Framework
fastapi
uvicorn[standard]
pydantic

# HTTP Client
requests
httpx
urllib3

# AI API Clients
anthropic
openai

# LangGraph for Agent Workflow
langgraph
langchain-core

# YAML Processing
pyyaml

# Utilities
python-multipart
python-dotenv

# Development & Testing
pytest
pytest-asyncio

# Treesitter for Code Parsing
tree_sitter
tree_sitter_python
tree_sitter_javascript
tree_sitter_typescript
tree_sitter_java
tree_sitter_cpp
tree_sitter_go
tree_sitter_ruby

# Packaging
pyinstaller